Updating the routing, controlling and model for the todos, so that they work in tandem with the users

+++ controllers/todo.js
@@ -1,9 +1,12 @@
 const Todo = require("../models/todo");
+const pool = require("../db.js");
 
 const getAllTodos = async (req, res) => {
 	try {
-		const todos = await Todo.getTodos();
-		res.json(todos);
+		const userID = await pool.query(`SELECT id FROM users WHERE username='${req.body.username}'`);
+		if (userID.rows.length == 0) return res.status(400).json({ error: `No such username: ${req.body.username}` });
+		const todos = await Todo.getTodos(userID.rows[0].id);
+		res.status(200).json(todos);
 	} catch (err) {
 		res.status(500).json({ error: err.message });
 	}
@@ -12,9 +15,11 @@ const getAllTodos = async (req, res) => {
 const getTodo = async (req, res) => {
 	try {
 		const { id } = req.params;
-		const todo = await Todo.getTodoById(id);
+		const userID = await pool.query(`SELECT id FROM users WHERE username='${req.body.username}'`);
+		if (userID.rows.length == 0) return res.status(400).json({ error: `No such username: ${req.body.username}` });
+		const todo = await Todo.getTodoById(id, userID.rows[0].id);
 		if (!todo) return res.status(404).json({ error: "Todo not found" });
-		res.json(todo);
+		res.status(200).json(todo);
 	} catch (err) {
 		res.status(500).json({ error: err.message });
 	}
@@ -23,10 +28,10 @@ const getTodo = async (req, res) => {
 const createNewTodo = async (req, res) => {
 	try {
 		const { task, username } = req.body;
-		const userID = await pool.query("SELECT id FROM users WHERE username=$1", [username]);
-		console.log(userID);
+		const userID = await pool.query(`SELECT id FROM users WHERE username='${username}'`);
+		if (userID.rows.length == 0) return res.status(400).json({ error: `No such username: ${username}` });
 		if (!task) return res.status(400).json({ error: "Task is required" });
-		const todo = await Todo.createTodo(task, username);
+		const todo = await Todo.createTodo(task, userID.rows[0].id);
 		res.status(201).json(todo);
 	} catch (err) {
 		res.status(500).json({ error: err.message });
@@ -36,9 +41,12 @@ const createNewTodo = async (req, res) => {
 const updateTodoStatus = async (req, res) => {
 	try {
 		const { id } = req.params;
+		const userID = await pool.query(`SELECT id FROM users WHERE username='${req.body.username}'`);
+		if (userID.rows.length == 0) return res.status(400).json({ error: `No such username: ${req.body.username}` });
+
 		const updatedTodo = await Todo.updateTodo(id);
 		if (!updatedTodo) return res.status(404).json({ error: "Todo not found" });
-		res.json(updatedTodo);
+		res.status(200).json(updatedTodo);
 	} catch (err) {
 		res.status(500).json({ error: err.message });
 	}
@@ -47,8 +55,10 @@ const updateTodoStatus = async (req, res) => {
 const deleteTodoById = async (req, res) => {
 	try {
 		const { id } = req.params;
+		const userID = await pool.query(`SELECT id FROM users WHERE username='${req.body.username}'`);
+		if (userID.rows.length == 0) return res.status(400).json({ error: `No such username: ${req.body.username}` });
 		await Todo.deleteTodo(id);
-		res.json({ message: "Todo deleted successfully" });
+		res.status(204).json({ message: "Todo deleted successfully" });
 	} catch (err) {
 		res.status(500).json({ error: err.message });
 	}
